+++ pages/_app.js
@@ -1,21 +1,12 @@
 import GlobalStyle from "@/styles";
 import Head from "next/head";
 import { SWRConfig } from "swr";
+import fetcher from "@/utilities/fetcher";
 
 export default function App({ Component, pageProps }) {
   return (
     <>
-      <SWRConfig
-        value={{
-          fetcher: async (...args) => {
-            const response = await fetch(...args);
-            if (!response.ok) {
-              throw new Error(`Request with ${JSON.stringify(args)} failed.`);
-            }
-            return await response.json();
-          },
-        }}
-      >
+      <SWRConfig value={{ fetcher }}>
         <GlobalStyle />
         <Head>
           <title>Team Master</title>
