+++ utilities/fetcher.js
@@ -1,0 +1,7 @@
+export default async function fetcher(...args) {
+  const response = await fetch(...args);
+  if (!response.ok) {
+    throw new Error(`Request with ${JSON.stringify(args)} failed.`);
+  }
+  return await response.json();
+}
